+++ velox/duckdb/README.md
@@ -9,13 +9,11 @@ a new feature, first clone DuckDB git repository:
 Then generate the amalgamated .cpp and .hpp files:
 
     python3 scripts/amalgamation.py --extended --splits=8
-    python3 scripts/parquet_amalgamation.py
 
 Then copy the generated files to velox/external/duckdb:
 
     export VELOX_PATH="<path/to/velox>"
     rsync -vrh src/amalgamation/duckdb* ${VELOX_PATH}/velox/external/duckdb/
-    rsync -vrh src/amalgamation/parquet* ${VELOX_PATH}/velox/external/duckdb/
 
 We also maintain a copy of TPC-H dataset generators that need to be updated:
